Guiler tek paket halinde toparlandı.
Artık Main classı gui değil.Sadece başlatıcı class olarak görev
yapıyor.İlk olarak browse guiyi çağırıyor.
