+++ nbproject/private/private.xml
@@ -3,8 +3,9 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/Harun/Documents/NetBeansProjects/Nesneye%20Dayalı%20Programlama/test/com/hns/oop/article/ArticleTest.java</file>
-            <file>file:/C:/Users/Harun/Documents/NetBeansProjects/Nesneye%20Dayalı%20Programlama/src/com/hns/oop/article/ArticleDatabase.java</file>
+            <file>file:/C:/Projeler/Nesneye-Dayali-Programlama/src/com/hns/oop/Main.java</file>
+            <file>file:/C:/Projeler/Nesneye-Dayali-Programlama/src/com/hns/oop/article/ArticleDatabase.java</file>
+            <file>file:/C:/Projeler/Nesneye-Dayali-Programlama/src/com/hns/oop/Downloader.java</file>
         </group>
     </open-files>
 </project-private>
